split settings components and enforce dynamic to pass build
