+++ src/app/admin/settings/capacity-card.tsx
@@ -1,11 +1,5 @@
-"use client";
-
-import { useMutation, useSuspenseQuery } from "@tanstack/react-query";
-import { Loader2 } from "lucide-react";
 import { Suspense } from "react";
-import { toast } from "sonner";
 import { PendingFormData } from "@/components/form/pending-form-data";
-import { useSettingsForm } from "@/components/form/use-settings-form";
 import {
   Card,
   CardContent,
@@ -13,8 +7,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { useTRPC } from "@/lib/trpc/react";
-import { SSettings } from "@/server/db/types";
+import { CapacityForm } from "./capacity-form";
 
 export function CapacityCard() {
   return (
@@ -29,70 +22,9 @@ export function CapacityCard() {
 
       <CardContent>
         <Suspense fallback={<PendingFormData />}>
-          <CapacityCardForm />
+          <CapacityForm />
         </Suspense>
       </CardContent>
     </Card>
-  );
-}
-
-function CapacityCardForm() {
-  const trpc = useTRPC();
-  const { data } = useSuspenseQuery(trpc.settings.get.queryOptions());
-  const { mutate, isPending } = useMutation(
-    trpc.settings.patch.mutationOptions({
-      onSuccess: () => toast.success("Paramètres mis à jour"),
-      onError: (e) => toast.error(e.message),
-    }),
-  );
-  const form = useSettingsForm({
-    defaultValues: data,
-    onSubmit: ({ value }) => {
-      mutate(value);
-    },
-    validators: {
-      onSubmit: SSettings,
-    },
-  });
-  return (
-    <form
-      className="space-y-4"
-      autoComplete="on"
-      noValidate
-      onSubmit={(e) => {
-        e.preventDefault();
-      }}
-    >
-      <form.AppField name="maxGuestsPerBooking">
-        {(field) => (
-          <field.NumberField
-            id="maxGuestsPerBooking"
-            label="Nombre maximum d'invités pour une réservation"
-          />
-        )}
-      </form.AppField>
-      <form.AppField name="maxCapacityPerSlot">
-        {(field) => (
-          <field.NumberField
-            id="maxCapacityPerSlot"
-            label=" Capacité maximum par créneau"
-          />
-        )}
-      </form.AppField>
-      <form.AppField name="maxCapacityPerService">
-        {(field) => (
-          <field.NumberField
-            id="maxCapacityPerSlot"
-            label=" Capacité maximum par service"
-          />
-        )}
-      </form.AppField>
-      <form.AppForm>
-        <form.SubmitButton>
-          {isPending && <Loader2 className="animate-spin" />}
-          Sauvegarder
-        </form.SubmitButton>
-      </form.AppForm>
-    </form>
   );
 }
