+++ src/app/admin/settings/exceptions-card.tsx
@@ -1,20 +1,12 @@
-"use client";
-
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { ChevronDown, ChevronRight } from "lucide-react";
-import { Suspense, useState } from "react";
+import { Suspense } from "react";
 import { PendingFormData } from "@/components/form/pending-form-data";
-import { Button } from "@/components/ui/button";
 import {
   Card,
   CardDescription,
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Table, TableBody, TableCell, TableRow } from "@/components/ui/table";
-import { useTRPC } from "@/lib/trpc/react";
-import { DateFormat } from "@/lib/utils";
-import { SubTable } from "./weekly-card";
+import { ExceptionsTable } from "./exceptions-table";
 
 export function ExceptionsCard() {
   return (
@@ -40,71 +32,5 @@ export function ExceptionsCard() {
         <ExceptionsTable />
       </Suspense>
     </Card>
-  );
-}
-
-function ExceptionsTable() {
-  const trpc = useTRPC();
-  const { data } = useSuspenseQuery(trpc.schedule.getExceptions.queryOptions());
-  const [closedDays, setClosedDays] = useState<Record<number, boolean>>({});
-  const toggleDay = (day: number) => {
-    setClosedDays((prev) => ({ ...prev, [day]: !prev[day] }));
-  };
-  const hasExceptions = data.length > 0;
-  return (
-    <Table>
-      <TableBody>
-        {hasExceptions ? (
-          data.map((e, i) => {
-            const hasPeriods = e.periods.length > 0;
-            return (
-              <>
-                <TableRow key={crypto.randomUUID()} className="bg-muted/50">
-                  <TableCell>
-                    <div className="flex justify-between items-center">
-                      {hasPeriods ? (
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          onClick={() => toggleDay(i)}
-                        >
-                          {closedDays[i] ? (
-                            <ChevronRight className="size-4" />
-                          ) : (
-                            <ChevronDown className="size-4" />
-                          )}
-                        </Button>
-                      ) : (
-                        <div />
-                      )}
-                      <div className="font-bold">
-                        Du {DateFormat.format(e.from)}
-                        {e.to && ` au ${DateFormat.format(e.to)}`}
-                        {!hasPeriods && " (fermé)"}
-                      </div>
-                      <div />
-                      {/* <Button variant="ghost" size="icon">
-                        <Plus className="size-4" />
-                      </Button> */}
-                    </div>
-                  </TableCell>
-                </TableRow>
-                {!closedDays[i] && hasPeriods && (
-                  <TableRow>
-                    <TableCell colSpan={3}>
-                      <SubTable periods={e.periods} />
-                    </TableCell>
-                  </TableRow>
-                )}
-              </>
-            );
-          })
-        ) : (
-          <TableRow>
-            <TableCell className="text-center">Aucune exception</TableCell>
-          </TableRow>
-        )}
-      </TableBody>
-    </Table>
   );
 }
