+++ src/app/admin/settings/page.tsx
@@ -1,8 +1,10 @@
 import { CapacityCard } from "./capacity-card";
 import { ExceptionsCard } from "./exceptions-card";
-import WeeklyCard from "./weekly-card";
+import { WeeklyCard } from "./weekly-card";
 
-export default async function SettingsPage() {
+export const dynamic = "force-dynamic";
+
+export default function SettingsPage() {
   return (
     <div className="flex gap-3 items-start">
       <WeeklyCard />
